Implement SymbolicHamiltonian.__matmul__
